ADD - add a setting to enable/disable the delete action on every location

+++ displayPlaceInfos.php
@@ -33,8 +33,10 @@ if(isset($_GET['id'])){
     <?php endif; ?>
     <div class="description" id="desc'+selectedFeature.get('id')+'"><?=utf8_encode(nl2br($stripslashes($row['description'])))?></div>
     <?php if($settings->getSettingValue('cbmp_application_EditLocation')=='enabled') : ?>
-    <input type="button" value="Edit" onclick="getEditForm(this,<?=$id?>);" >
-    <input type="button" value="Delete" onclick="deletePlace('<?=htmlspecialchars(addslashes($stripslashes(utf8_encode($row['placeName']))))?>',<?=$id?>);"/>
+        <input type="button" value="Edit" onclick="getEditForm(this,<?=$id?>);" >
+        <?php if($settings->getSettingValue('cbmp_application_DeleteLocation')=='enabled') : ?>
+            <input type="button" value="Delete" onclick="deletePlace('<?=htmlspecialchars(addslashes($stripslashes(utf8_encode($row['placeName']))))?>',<?=$id?>);"/>
+        <?php endif; ?>
     <?php endif; ?>
 </div>
 <?php
